pyproject.toml
[project]
name = "google-play-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.1.2",
    "google-api-python-client>=2.167.0",
    "google-auth>=2.39.0",
]
